+++ dev/Network.js
@@ -2,17 +2,26 @@
 
 class Network {
 
-    constructor ({learningRate=0.2, layers=[], adaptiveLR="noAdaptiveLR", activation="sigmoid", cost="crossEntropy"}={}) {
+    constructor ({learningRate, layers=[], adaptiveLR="noAdaptiveLR", activation="sigmoid", cost="crossEntropy", rmsDecay}={}) {
         this.state = "not-defined"
         this.layers = []
         this.epochs = 0
         this.iterations = 0
 
-        this.learningRate = learningRate
+        if(learningRate!=undefined && learningRate!=null) {
+            this.learningRate = learningRate
+        }else {
+            this.learningRate = adaptiveLR=="RMSProp" ? 0.01 : 0.2
+        }
+
         this.adaptiveLR = [false, null, undefined].includes(adaptiveLR) ? "noAdaptiveLR" : adaptiveLR
         this.weightUpdateFn = NetMath[this.adaptiveLR]
         this.activation = NetMath[activation]
         this.cost = NetMath[cost]
+
+        if(this.adaptiveLR=="RMSProp"){
+            this.rmsDecay = rmsDecay==undefined ? 0.99 : rmsDecay
+        }
 
         if(layers.length) {
 
@@ -203,7 +212,7 @@ class Network {
 
             const testInput = () => {
 
-                console.log("Testing iteration", testIteration+1, totalError/(testIteration+1)/100)
+                console.log("Testing iteration", testIteration+1, totalError/(testIteration+1))
 
                 const output = this.forward(testSet[testIteration].input)
                 const target = testSet[testIteration].expected || testSet[testIteration].output
@@ -214,7 +223,7 @@ class Network {
 
                 if(testIteration < testSet.length)
                     setTimeout(testInput.bind(this), 0)
-                else resolve(totalError/testSet.length/100)
+                else resolve(totalError/testSet.length)
             }
             testInput()
         })