+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # Upcoming
 ---
 #### Network
-- Added Adagrad as adaptiveLR configuration
-- Added gain as adaptiveLR configuration
 - Added adaptiveLR as a network configuration
 #### NetMath
+- Added RMSProp as adaptiveLR configuration
+- Added Adagrad as adaptiveLR configuration
+- Added gain as adaptiveLR configuration
 - Added Mean Squared Error cost function
-
 # v1.0.0
 ----
 Initial release